requirements.txt
pypiwin32
pyyaml
requests
twitchAPI